requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
Flask-Compress==1.14
Flask-Caching==2.1.0
Flask-Mail==0.9.1
WTForms==3.1.1
Werkzeug==3.0.1
python-dotenv==1.0.0
Pillow==10.0.1
email-validator==2.1.0
itsdangerous==2.1.2
